cnn/pyproject.toml
[project]
name = "cnn"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "torch",
    "torchvision",
    "tqdm>=4.67.1",
    "pillow",
    "numpy",
]
